backend/requirements.txt
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
drf-yasg==1.17.1
requests==2.24.0
gunicorn==20.0.4